+++ lambda/strings.c
@@ -11,8 +11,8 @@
 #include <string.h>
 
 void split(char* const string, char* const delim,
-		char* fields[], uint8_t const size) {
-	uint8_t index = 0;
+		char* fields[], size_t const size) {
+	size_t index = 0;
 	char* token = strtok(string, delim);
 	while (token != NULL && index < size) {
 		fields[index++] = token;
